+++ CHANGELOG.md
@@ -13,9 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [projects] New progress modal showing realtime creation progress
 
 ### Changed
-* [database] Create and list endpoints now return objects rather than just names
+* [database] Create and list endpoints return DatabaseData objects, not just names
 
 ### Fixed
+* [database] Tests no longer create databases using the real connection
 * [projects] Removed a rogue '>' from the New Project confirmation page
 * [projects] Successful project creation no longer prompts to discard changes